relationship refactoring
