+++ Sources/SwiftletModel/Relationship/Relation+Conveniece.swift
@@ -7,6 +7,8 @@
 
 import Foundation
 
+//MARK: - Relation Extensions
+
 public extension Relation {
     static var none: Self {
         Relation(state: .none)
@@ -58,5 +60,62 @@ public extension Relation where Cardinality == Relations.ToMany<Entity> {
  
     static func appending(ids: [Entity.ID]) -> Self {
         Relation(state: State(ids: ids, slice: true))
+    }
+}
+
+//MARK: - Relationship Extensions
+
+public extension Relationship where Cardinality == Relations.ToOne<Entity> {
+
+    static func id(_ id: Entity.ID) -> Self {
+        Relationship(relation: .id(id))
+    }
+
+    static func relation(_ entity: Entity) -> Self {
+        Relationship(relation: .relation(entity))
+    }
+    
+    static func fragment(_ entity: Entity) -> Self {
+        Relationship(relation: .fragment( entity))
+    }
+}
+
+public extension Relationship where Cardinality == Relations.ToMany<Entity> {
+
+    static func ids(_ ids: [Entity.ID]) -> Self {
+        Relationship(relation: .ids(ids))
+    }
+
+    static func relation(_ entities: [Entity]) -> Self {
+        Relationship(relation: .relation(entities))
+    }
+    
+    static func fragment(_ entities: [Entity]) -> Self {
+        Relationship(relation: .fragment(entities))
+    }
+
+    static func appending(ids: [Entity.ID]) -> Self {
+        Relationship(relation: .appending(ids: ids))
+    }
+
+    static func appending(relation entities: [Entity]) -> Self {
+        Relationship(relation: .appending(relation: entities))
+    }
+    
+    static func appending(fragment entities: [Entity]) -> Self {
+        Relationship(relation: .appending(fragment: entities))
+    }
+}
+
+public extension Relationship where Cardinality == Relations.ToOne<Entity>, Constraints: OptionalRelation {
+    static var null: Self {
+        Relationship(relation: .null)
+    }
+}
+
+public extension Relationship {
+
+    static var none: Self {
+        Relationship(relation: .none)
     }
 }
